requirements.txt
bison==0.1.2
cairocffi==1.2.0
cffi==1.14.5
cmake==3.18.4.post1
pkg-resources==0.0.0
pycparser==2.20
python-igraph==0.9.0
PyYAML==5.4.1
texttable==1.6.3
